refactor(8bc): decompose AdminLocationsScreen + NotificationsScreen, complete Bundle 8
Bundle 8B — AdminLocationsScreen (807 -> 451 lines):
- hooks/use-confirm.ts: extracted inlined useConfirmLocal promise-confirm pattern
- components/admin/locations/AddLocationForm.tsx: name input + Add/cancel
- components/admin/locations/DangerZone.tsx: Archive/Delete section (~310 lines
  of JSX) — owns UI state + API calls, delegates list reconciliation to parent
  via onArchived/onDeleted callbacks (preserves the original patch-vs-refetch
  logic exactly)

Bundle 8C — NotificationsScreen (1234 -> 651 lines):
- utils/notifications.ts: pure helpers (relativeTime, formatBookingDate,
  urlBase64ToUint8Array, arrayBufferToBase64) + constants (PAGE_SIZE,
  PIN_STORAGE_KEY, TYPE_LABELS, TYPE_ICONS, TYPE_FILTERS)
- components/admin/notifications/PushBanner.tsx: web push opt-in banner +
  usePushStatus hook (moved wholesale, was already structurally isolated)
- components/admin/notifications/NotificationRow.tsx: the ~150-line row
  renderer (type icon, title/meta, pin/read/delete actions, swipe wrapper)
- components/admin/notifications/notifications.styles.ts: inline styles moved

Tests: +24 net new (4 useConfirm, 5 AddLocationForm, 5 NotificationRow,
10 utils/notifications). The 75 existing screen-level tests (locations 30 +
notifications 45) pass unchanged — they stub child components and assert the
screen's contract, so decomposition is invisible to them.

Frontend 1587 pass / 0 fail (was 1563). tsc + prettier + lint clean.
Bundle 8 complete — unblocks Bundles 9, 10, 11, 12.
